Remove dependence on prior_cluster.py and m2_config. Add inheritance class GalaxyClusterPrior. Support gaussian galaxy shape and identical r-band cluster.
