--- case_studies/galaxy_clustering/m2_config.yaml
@@ -1,82 +1,0 @@
----
-defaults:
-    - ../../bliss/conf@_here_: base_config
-    - _self_
-    - override hydra/job_logging: stdout
-
-encoder:
-    tile_slen: 2
-    min_flux_threshold: 0.9419  # (1.59 is 22 mag; 0.63 is 23 mag; 0.25 is 24 mag)
-    double_detect: true
-    tiles_to_crop: 1
-    image_normalizer:
-        bands: [2]
-    matcher:
-        _target_: bliss.encoder.metrics.CatalogMatcher
-        dist_slack: 0.5
-        mag_slack: 0.5
-        mag_band: 2  # SDSS r-band
-    metrics:
-        _target_: torchmetrics.MetricCollection
-        metrics:
-          - _target_: bliss.encoder.metrics.DetectionPerformance
-            mag_bin_cutoffs: [19, 19.4, 19.8, 20.2, 20.6, 21, 21.4, 21.8, 22.065]
-            exclude_last_bin: true
-
-prior:
-    _target_: bliss.simulator.prior_cluster.ClusterPrior
-    survey_bands: ["u", "g", "r", "i", "z"]  # SDSS available band filters
-    reference_band: 2  # SDSS r-band
-    star_color_model_path: ${simulator.survey.dir_path}/color_models/star_gmm_nmgy.pkl
-    gal_color_model_path: ${simulator.survey.dir_path}/color_models/gal_gmm_nmgy.pkl
-    n_tiles_h: 56
-    n_tiles_w: 56
-    tile_slen: 2
-    batch_size: 2
-    max_sources: 6
-    mean_sources: 0.48
-    min_sources: 0
-    prob_galaxy: 0.0
-    star_flux_exponent: 0.9859821185389767
-    star_flux_truncation: 5685.588160703261
-    star_flux_loc: -1.162430157551662
-    star_flux_scale: 1.4137911256506595
-
-surveys:
-    sdss:
-        # need to override this for proper data generation
-        fields:
-            - run: 2583
-              camcol: 2
-              fields: [136]
-        background_offset: 80
-
-generate:
-    n_image_files: 16
-    n_batches_per_file: 16
-    cached_data_path: /data/scratch/regier/m2
-
-cached_simulator:
-    cached_data_path: /data/scratch/regier/m2
-    batch_size: 23
-
-train:
-    trainer:
-        precision: 16-mixed
-
-predict:
-    dataset:
-        _target_: bliss.surveys.sdss.SloanDigitalSkySurvey
-        dir_path: ${paths.sdss}
-        fields:
-            - run: 2583
-              camcol: 2
-              fields: [136]
-        psf_config:
-            pixel_scale: 0.396
-            psf_slen: 25
-        pixel_shift: 2
-        align_to_band: null
-        load_image_data: true
-        crop_config: [624, 736, 304, 416]
-    weight_save_path: null
